Move WAP3 to the right directory
